Image classification model trainings/MobileNetV3/requirements.txt
Import the following:

torch
torchvision
pillow
pandas
tqdm
opencv-python
matplotlib
scikit-learn